корректировка классов, реализация пост методов

+++ src/main/java/com/example/service/UserServiceImpl.java
@@ -10,6 +10,7 @@ import com.example.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,10 +26,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<UserDtoOutput> getAllUsers(List<Long> ids, final int from, final int size) {
-        PageRequest pageRequest = PageRequest.of(from / size, size);
+    public List<UserDtoOutput> getAllUsers(List<Long> ids, final int page, final int size) {
+        PageRequest pageRequest = PageRequest.of(page, Math.min(size, 50), Sort.by(Sort.Direction.DESC, "createdAt"));
         log.info("Запрос на получение списка пользователей.");
         final List<User> users;
+
         if (Objects.isNull(ids) || ids.isEmpty()) {
             users = userRepository.findAll(pageRequest).getContent();
             log.info("Получен список всех пользователей.");
@@ -36,17 +38,17 @@ public class UserServiceImpl implements UserService {
             users = userRepository.findByIdIn(ids, pageRequest);
             log.info("Получен список пользователей по заданным id.");
         }
-        return UserMapper.toListDto(users);
 
+        return UserMapper.toListDto(users);
     }
 
     @Override
     @Transactional(readOnly = true)
     public UserDtoOutput getUserById(final long userId) {
         final User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundException("Пользователь с id = {} не найден." + userId));
+                .orElseThrow(() -> new NotFoundException(String.format("Пользователь с id = %d не найден.", userId)));
 
-        log.info("Получение пользователя по id = {}.", userId);
+        log.info("Получен пользователь с id = {}.", userId);
         return UserMapper.toUserDto(user);
     }
 
@@ -56,14 +58,19 @@ public class UserServiceImpl implements UserService {
             log.warn("Пользователь уже существует.");
             throw new DuplicatedDataException("Пользователь уже существует.");
         }
-
         final User user = userRepository.save(UserMapper.toUser(userDtoInput));
+
         log.info("Пользователь с id = {} добавлен.", user.getId());
         return UserMapper.toUserDto(user);
     }
 
     @Override
-    public void deleteUser(final Long userId) {
+    public void deleteUser(final long userId) {
+        if (!userRepository.existsById(userId)) {
+            log.warn("Пост для удаления не найден: id = {}.", userId);
+            throw new NotFoundException(String.format("Пользователь с id = %d не найден.", userId));
+        }
+
         userRepository.deleteById(userId);
         log.info("Пользователь с id  = {} удален.", userId);
     }
